% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_zips}
\alias{nyc_zips}
\title{New York City zip codes}
\format{A simple feature collection with 262 features (rows) and 11
    fields (12 columns).
\describe{
\item{\code{objectid}}{Object identifier.}
\item{\code{zip_code}}{Zip code.}
\item{\code{po_name}}{Post office region name.}
\item{\code{state}}{State (all NY).}
\item{\code{borough}}{City borough.}
\item{\code{st_fips}}{State FIPS code (all 36).}
\item{\code{cty_fips}}{County FIPS code.}
\item{\code{bld_gpostal_code}}{Building postal code.}
\item{\code{shape_leng}}{Shape length.}
\item{\code{shape_area}}{Shape area}
\item{\code{vstrat}}{Variance stratum}
\item{\code{x_id}}{x_id}
\item{\code{geometry}}{POLYGON [°] description of the shape of the
    zip code area.}
}}
\source{
.
}
\usage{
nyc_zips
}
\description{
A simple features object containing polygons for the 262 zip code
areas in New York city, along with additional geographic information.
}
\keyword{datasets}
