% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycdogs-package.R
\docType{package}
\name{nycdogs-package}
\alias{nycdogs}
\alias{nycdogs-package}
\title{nycdogs: New York City Dog Licenses and Bite Incidents by Zip Code}
\description{
A data package containing two datasets on dogs in New York City: active dog licenses from the Department of Health and Mental Hygiene, and reported dog bite incidents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kjhealy.github.io/nycdogs/}
  \item \url{https://github.com/kjhealy/nycdogs}
  \item Report bugs at \url{https://github.com/kjhealy/nycdogs/issues}
}

}
\author{
\strong{Maintainer}: Kieran Healy \email{kjhealy@gmail.com} (\href{https://orcid.org/0000-0001-9114-981X}{ORCID})

Authors:
\itemize{
  \item Kieran Healy \email{kjhealy@gmail.com} (\href{https://orcid.org/0000-0001-9114-981X}{ORCID})
}

}
\keyword{internal}
