% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_bites}
\alias{nyc_bites}
\title{New York City Dog Reported Dog bites}
\format{
A tibble containing 10,280 rows and 11 columns
\describe{
\item{\code{uniqueid}}{Incident identified.}
\item{\code{date_of_bite}}{Date of bite.}
\item{\code{species}}{Species of animal (all "Dog").}
\item{\code{breed}}{Breed of dog.}
\item{\code{age}}{Age of dog.}
\item{\code{gender}}{Sex of dog.}
\item{\code{spay_neuter}}{Spayed or neutered (True/False)}
\item{\code{borough}}{Borough where bite occurred.}
\item{\code{zip_code}}{Zip code where bite occurred.}
\item{\code{year}}{Year}
\item{\code{breed_rc}}{Recoded breed variable with aggregated breed categories)}
}
}
\source{
New York City public data
}
\usage{
nyc_bites
}
\description{
Reported dog bite incidents in New York city between January 1st 2015
and Decenber 31st 2017,
}
\keyword{datasets}
