% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_license}
\alias{nyc_license}
\title{Licensed Dogs in New York City as of January 1st 2017}
\format{A tibble containing 122,203 rows and 14 columns
\describe{
\item{\code{animal_name}}{Name of dog.}
\item{\code{animal_gender}}{Sex of dog.}
\item{\code{animal_birth_month}}{Birth month.}
\item{\code{breed_rc}}{Breed of dog (simplier categories).}
\item{\code{borough}}{Borough of owner.}
\item{\code{zip_code}}{Zip code of owner.}
\item{\code{community_district}}{Community district of owner}
\item{\code{nta}}{Neighborhood Tabulation Area.}
\item{\code{city_council_district}}{City council district.}
\item{\code{congressional_district}}{Congressional district}
\item{\code{state_senatorial_district}}{State Senatorial district}
\item{\code{license_issued_date}}{Date license issued}
\item{\code{license_expired_date}}{Date license expired}
}}
\source{
New York City Public Data
}
\usage{
nyc_license
}
\description{
Dogs in New York city with licenses issued between
September 12th 2014 and December 31st 2016.
}
\keyword{datasets}
